part1: I added lines for the third to fifth values and added those values to the equation in line 27.
part2: I added an "int" in front of the input in line 13.

part3: I fixed the 6 errors in the code, one of which was adding parentheses.

part4: I wrote a program that calculates the area of a triangle.

part5: I wrote a program that prints the user's first and last name.